Move safe reader to libbeat/common/encoding/xml package
